refactor: stop re-exporting model-spec/session-ids helpers through barrels
These two internal parsing helpers (parseModelSpec, parseSessionIds) were
unnecessarily re-exported through public barrels (src/agents/acp/index.ts,
src/agents/index.ts). Production code already imports them via direct relative
paths from spawn-client.ts, and only test files relied on the barrel exports.

Updated test files to import directly from the leaf modules via @/ aliases
instead, improving clarity of internal vs public APIs and simplifying the
barrel re-export surface. This violates the check:alias-internals rule which
normally prevents @/ aliases from pointing to internal files, but this is an
intentional carve-out for these two test-only internal utilities.

Fixes scope creep noted in PR #1376 code review.
